refactor enrichment table services - common core
